+++ money-calculator/src/main/java/software/ulpgc/moneycalculator/Model/Money.java
@@ -2,7 +2,7 @@ package software.ulpgc.moneycalculator.Model;
 
 import software.ulpgc.moneycalculator.Model.Currency;
 
-public record Money(long amount, Currency currency) {
+public record Money(double amount, Currency currency) {
     @Override
     public String toString() {
         return amount + " " + currency;
